refactor(cleanup): break CLI check barrel and run↔onboard cycles (pass 11)
Extracted the CheckResult interface from cli/src/checks/index.ts into a
new cli/src/checks/types.ts. The 9 individual check files now import
CheckResult from ./types.js instead of from the barrel, eliminating the
9 cycles where each check imported a type from the index that re-
exported the check back.

Also converted cli/src/commands/run.ts to dynamic-import onboard.ts
inside runCommand. onboard.ts already dynamic-imports run.ts, so the
cycle is now two-way dynamic (runtime-safe because ESM dynamic imports
defer resolution, even though madge static analysis still reports the
edge).

The CheckResult interface is re-exported from the barrel index.ts, so
external consumers (cli/src/commands/doctor.ts) are unaffected.

Addresses .cleanup-reports/04-circular.md cycles 1-10. Running
`pnpm dlx madge --circular` across the full repo:
- Before: 16 cycles
- After: 7 cycles (cycle 10 remains on madge output but is now runtime-
  safe via bidirectional dynamic imports)

Remaining cycles are either:
- Bidirectional FK schemas (cycle 11, packages/db) — can't be fixed
  with type-only imports because Drizzle's deferred reference callback
  accesses runtime values on the imported table (executionWorkspaces.id)
- Plugin services (cycles 12-14) — intentional mutual-type pattern per
  report, type-only imports prevent runtime issues
- UI adapters (cycles 15-16) — type-only false positives

Typecheck: all 20 workspace packages green.
